test_requirements.txt
# Tests
assertionlib
codacy-coverage
pytest
pytest-cov
pytest-mock
Cython
setuptools
Nano-Utils>=2.0

# Runtime dependencies
h5py
mendeleev
more-itertools
noodles==0.3.3
numpy
scipy
schema
pyyaml>=5.1
plams>=1.5.1
qmflows>=0.11.2
